feat(view): delegate CustomerDashboard input handling to view layer

+++ src/com/ecommerce/controller/CartController.java
@@ -2,6 +2,7 @@ package com.ecommerce.controller;
 
 import com.ecommerce.exception.CartItemNotFoundException;
 import com.ecommerce.exception.InvalidProductException;
+import com.ecommerce.exception.InvalidQuantityException;
 import com.ecommerce.model.entities.Cart;
 import com.ecommerce.model.entities.CartItem;
 import com.ecommerce.model.entities.Customer;
@@ -9,9 +10,10 @@ import com.ecommerce.service.CartService;
 import com.ecommerce.view.CartView;
 
 import java.util.List;
+import java.util.logging.Logger;
 
 public class CartController {
-
+    private Logger logger = Logger.getLogger(CartController.class.getName());
     private final CartService cartService;
     private final CartView cartView;
 
@@ -21,14 +23,22 @@ public class CartController {
         this.cartView = cartView;
     }
 
-    public void addProductToCart(Customer customer, int productId, int quantity) {
+    public void addProductToCart(Customer customer) {
+
         try {
+            int productId = cartView.productIdPrompt();
+            if (productId == -1) {
+                return;
+            }
+            int quantity = cartView.quantityPrompt();
+
             cartService.addProductToCart(customer, productId, quantity);
             cartView.showSuccessMessage();
-        } catch (InvalidProductException e) {
+        } catch (NumberFormatException |InvalidQuantityException | InvalidProductException | CartItemNotFoundException e) {
             cartView.showErrorMessage(e.getMessage());
         }
     }
+
     public void getCartItems(Customer customer) {
         Cart cart = new Cart(customer.getId());
         List<CartItem> items = cartService.getLoadedItems(cart);
@@ -36,26 +46,57 @@ public class CartController {
     }
 
 
-    public void removeProductFromCart(Customer customer, int productId){
+    public void removeProductFromCart(Customer customer) {
 
-        try{
+        try {
+            int productId = cartView.productIdRemovePrompt();
+            if (productId == -1) {
+                return;
+            }
             cartService.removeItemFromCart(customer, productId);
+            logger.info("Successfully removed");
             cartView.showRemovedMessage();
 
-        }catch ( CartItemNotFoundException e){
+        } catch (CartItemNotFoundException e) {
+            logger.warning("No such item");
             cartView.showErrorMessage(e.getMessage());
         }
     }
 
-    public void updateQuantity(Customer customer, int productId, int newQuantity) {
-        cartService.updateCartItemQuantity(customer,productId, newQuantity);
+    public void UpdateQuantity(Customer customer) {
+        try {
+            int productId = cartView.productIdPrompt();
+            if (productId == -1) {
+                return;
+            }
+            int newQuantity = cartView.newQuantityPrompt();
+
+            cartService.updateCartItemQuantity(customer, productId, newQuantity);
+            cartView.showQuantityUpdated();
+        } catch (NumberFormatException e) {
+            System.out.println("❌ Invalid input. Please enter valid number");
+        } catch (InvalidQuantityException e) {
+            cartView.showErrorMessage(e.getMessage());
+        } catch (CartItemNotFoundException e) {
+            System.out.println("! " + e.getMessage());
+        } catch (Exception e) {
+            System.out.println("❌ Unexpected error: " + e.getMessage());
+        }
+
     }
 
     public void calculatePrice(Customer customer) {
-        double totalPrice = cartService.calculateTotalPrice(customer);
-        cartView.showTotalCartPrice(totalPrice);
+        try {
+            double totalPrice = cartService.calculateTotalPrice(customer);
+            logger.info("Cart total calculated");
+            cartView.showTotalCartPrice(totalPrice);
+        } catch (CartItemNotFoundException e) {
+            logger.warning(e.getMessage());
+            cartView.showErrorMessage(e.getMessage());
+        }
     }
-
-
-
 }
+
+
+
+
